Removed support for Os::Mutex alias
