+++ Os/Mutex.hpp
@@ -1,45 +1,76 @@
 // ======================================================================
 // \title Os/Mutex.hpp
-// \brief public Os::Mutex interface and alias
-//
-// This header aggregates all definitions needed to use Os::Mutex:
-// the interface, the configured alias, and the concrete delegate type.
-//
-// WARNING — include order is load-bearing. Do not reorder.
-//
-// The dependency constraints are:
-//
-//   1. config/OsDelegateMutex.hpp (CFG) defines the Os::Mutex type alias by
-//      forward-declaring a link-time delegate (e.g. DelegateMutex) or directly
-//      aliasing a concrete implementation (e.g. Va416x0Os::MaskingMutex).
-//      Must not include Os OSAL headers (they aren't yet defined).
-//      Should only be included in Os/MutexInterface.hpp.
-//
-//   2. Os/MutexInterface.hpp (IF) includes CFG first (so the Os::Mutex alias is
-//      available), then defines MutexHandle and MutexInterface.
-//
-//   3. OS_MUTEX_HEADER (IMPL) is defined by CFG and points to the concrete
-//      implementation header. If using delegation, this points to
-//      Os/DelegateMutex.hpp. If using compile-time selection, it points
-//      directly to a platform-specific implementation (e.g. MaskingMutex.hpp).
-//
-// MutexInterface.hpp must precede OS_MUTEX_HEADER here,
-// and CFG must never include either of them (that would form a cycle).
+// \brief common definitions for Os::Mutex
 // ======================================================================
 #ifndef Os_Mutex_hpp
 #define Os_Mutex_hpp
 
-#include "Os/MutexInterface.hpp"
-
-// Validate that OS_MUTEX_HEADER was defined by config/OsDelegateMutex.hpp
-#ifndef OS_MUTEX_HEADER
-#error "OS_MUTEX_HEADER must be defined in config/OsDelegateMutex.hpp"
-#endif
-
-#include OS_MUTEX_HEADER
+#include <Fw/FPrimeBasicTypes.hpp>
+#include <Os/Os.hpp>
 
 namespace Os {
 
+struct MutexHandle {};
+
+class MutexInterface {
+  public:
+    enum Status {
+        OP_OK,           //!< Operation was successful
+        ERROR_BUSY,      //!< Mutex is busy
+        ERROR_DEADLOCK,  //!< Deadlock condition detected
+        NOT_SUPPORTED,   //!< Mutex does not support operation
+        ERROR_OTHER      //!< All other errors
+    };
+
+    //! \brief default constructor
+    MutexInterface() = default;
+
+    //! \brief default virtual destructor
+    virtual ~MutexInterface() = default;
+
+    //! \brief copy constructor is forbidden
+    MutexInterface(const MutexInterface& other) = delete;
+
+    //! \brief copy constructor is forbidden
+    MutexInterface(const MutexInterface* other) = delete;
+
+    //! \brief assignment operator is forbidden
+    MutexInterface& operator=(const MutexInterface& other) = delete;
+
+    //! \brief return the underlying mutex handle (implementation specific)
+    //! \return internal mutex handle representation
+    virtual MutexHandle* getHandle() = 0;
+
+    //! \brief provide a pointer to a Mutex delegate object
+    static MutexInterface* getDelegate(MutexHandleStorage& aligned_new_memory);
+
+    virtual Status take() = 0;     //!<  lock the mutex return status
+    virtual Status release() = 0;  //!<  unlock the mutex return status
+};
+
+class Mutex final : public MutexInterface {
+  public:
+    Mutex();         //!<  Constructor. Mutex is unlocked when created
+    ~Mutex() final;  //!<  Destructor
+
+    //! \brief return the underlying mutex handle (implementation specific)
+    //! \return internal mutex handle representation
+    MutexHandle* getHandle() override;
+
+    Status take() override;            //!<  lock the mutex and get return status
+    Status release() override;         //!<  unlock the mutex and get return status
+    void lock();                       //!<  lock the mutex and assert success
+    void unLock();                     //!<  unlock the mutex and assert success
+    void unlock() { this->unLock(); }  //!<  alias for unLock to meet BasicLockable requirements
+
+  private:
+    // This section is used to store the implementation-defined mutex handle. To Os::Mutex and fprime, this type is
+    // opaque and thus normal allocation cannot be done. Instead, we allow the implementor to store then handle in
+    // the byte-array here and set `handle` to that address for storage.
+    //
+    alignas(FW_HANDLE_ALIGNMENT) MutexHandleStorage m_handle_storage;  //!< Mutex handle storage
+    MutexInterface& m_delegate;                                        //!< Delegate for the real implementation
+};
 //! \brief locks a mutex within the current scope
 //!
 //! The scope lock will lock the associated mutex immediately and will ensure the mutex is unlock when the scope lock
